zcash_client_backend: Move `OrchardPoolTester` here from `zcash_client_sqlite`
